+++ core/src/avm2/class.rs
@@ -1,6 +1,6 @@
 //! AVM2 classes
 
-use crate::avm2::function::Method;
+use crate::avm2::method::Method;
 use crate::avm2::names::{Multiname, Namespace, QName};
 use crate::avm2::r#trait::{Trait, TraitKind};
 use crate::avm2::script::TranslationUnit;
