+++ core/src/avm2/function.rs
@@ -2,178 +2,25 @@
 
 use crate::avm2::activation::Activation;
 use crate::avm2::class::Class;
+use crate::avm2::method::{BytecodeMethod, Method, NativeMethod};
 use crate::avm2::names::{Namespace, QName};
 use crate::avm2::object::{Object, ObjectPtr, TObject};
 use crate::avm2::r#trait::Trait;
 use crate::avm2::return_value::ReturnValue;
 use crate::avm2::scope::Scope;
-use crate::avm2::script::TranslationUnit;
 use crate::avm2::script_object::{ScriptObjectClass, ScriptObjectData};
 use crate::avm2::value::Value;
 use crate::avm2::{Avm2, Error};
 use crate::context::UpdateContext;
 use gc_arena::{Collect, CollectionContext, GcCell, MutationContext};
 use std::fmt;
-use std::rc::Rc;
-use swf::avm2::types::{AbcFile, Index, Method as AbcMethod, MethodBody as AbcMethodBody};
-
-#[derive(Clone, Debug, Collect)]
-#[collect(require_static)]
-pub struct CollectWrapper<T>(T);
-
-/// Represents a function defined in Ruffle's code.
-///
-/// Parameters are as follows:
-///
-///  * The AVM2 runtime
-///  * The action context
-///  * The current `this` object
-///  * The arguments this function was called with
-///
-/// Native functions are allowed to return a value or `None`. `None` indicates
-/// that the given value will not be returned on the stack and instead will
-/// resolve on the AVM stack, as if you had called a non-native function. If
-/// your function yields `None`, you must ensure that the top-most activation
-/// in the AVM1 runtime will return with the value of this function.
-pub type NativeFunction<'gc> = fn(
-    &mut Avm2<'gc>,
-    &mut UpdateContext<'_, 'gc, '_>,
-    Option<Object<'gc>>,
-    &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error>;
-
-/// Represents a reference to an AVM2 method and body.
-#[derive(Collect, Clone, Debug)]
-#[collect(no_drop)]
-pub struct Avm2MethodEntry<'gc> {
-    /// The translation unit this function was defined in.
-    pub txunit: TranslationUnit<'gc>,
-
-    /// The underlying ABC file of the above translation unit.
-    pub abc: CollectWrapper<Rc<AbcFile>>,
-
-    /// The ABC method this function uses.
-    pub abc_method: u32,
-
-    /// The ABC method body this function uses.
-    pub abc_method_body: u32,
-}
-
-impl<'gc> Avm2MethodEntry<'gc> {
-    /// Construct an `Avm2MethodEntry` from an `AbcFile` and method index.
-    ///
-    /// The method body index will be determined by searching through the ABC
-    /// for a matching method. If none exists, this function returns `None`.
-    pub fn from_method_index(
-        txunit: TranslationUnit<'gc>,
-        abc_method: Index<AbcMethod>,
-    ) -> Option<Self> {
-        let abc = txunit.abc();
-
-        if abc.methods.get(abc_method.0 as usize).is_some() {
-            for (index, method_body) in abc.method_bodies.iter().enumerate() {
-                if method_body.method.0 == abc_method.0 {
-                    return Some(Self {
-                        txunit,
-                        abc: CollectWrapper(txunit.abc()),
-                        abc_method: abc_method.0,
-                        abc_method_body: index as u32,
-                    });
-                }
-            }
-        }
-
-        None
-    }
-
-    /// Get the underlying ABC file.
-    #[allow(dead_code)]
-    pub fn abc(&self) -> Rc<AbcFile> {
-        self.txunit.abc()
-    }
-
-    /// Get the underlying translation unit this method was defined in.
-    pub fn translation_unit(&self) -> TranslationUnit<'gc> {
-        self.txunit
-    }
-
-    /// Get a reference to the ABC method entry this refers to.
-    pub fn method(&self) -> &AbcMethod {
-        &self.abc.0.methods.get(self.abc_method as usize).unwrap()
-    }
-
-    /// Get a reference to the ABC method body entry this refers to.
-    pub fn body(&self) -> &AbcMethodBody {
-        &self
-            .abc
-            .0
-            .method_bodies
-            .get(self.abc_method_body as usize)
-            .unwrap()
-    }
-}
-
-/// An uninstantiated method that can either be natively implemented or sourced
-/// from an ABC file.
-#[derive(Clone)]
-pub enum Method<'gc> {
-    /// A native method.
-    Native(NativeFunction<'gc>),
-
-    /// An ABC-provided method entry.
-    Entry(Avm2MethodEntry<'gc>),
-}
-
-unsafe impl<'gc> Collect for Method<'gc> {
-    fn trace(&self, cc: CollectionContext) {
-        match self {
-            Method::Native(_nf) => {}
-            Method::Entry(a2me) => a2me.trace(cc),
-        }
-    }
-}
-
-impl<'gc> fmt::Debug for Method<'gc> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Method::Native(_nf) => f
-                .debug_tuple("Method::Native")
-                .field(&"<native code>".to_string())
-                .finish(),
-            Method::Entry(a2me) => f.debug_tuple("Method::Entry").field(a2me).finish(),
-        }
-    }
-}
-
-impl<'gc> From<NativeFunction<'gc>> for Method<'gc> {
-    fn from(nf: NativeFunction<'gc>) -> Self {
-        Self::Native(nf)
-    }
-}
-
-impl<'gc> From<Avm2MethodEntry<'gc>> for Method<'gc> {
-    fn from(a2me: Avm2MethodEntry<'gc>) -> Self {
-        Self::Entry(a2me)
-    }
-}
-
-impl<'gc> Method<'gc> {
-    pub fn into_entry(self) -> Result<Avm2MethodEntry<'gc>, Error> {
-        match self {
-            Method::Native(_) => {
-                Err("Attempted to unwrap a native method as a user-defined one".into())
-            }
-            Method::Entry(a2me) => Ok(a2me),
-        }
-    }
-}
 
 /// Represents an AVM2 function.
 #[derive(Collect, Clone, Debug)]
 #[collect(no_drop)]
 pub struct Avm2Function<'gc> {
     /// The AVM method entry used to create this function.
-    pub method: Avm2MethodEntry<'gc>,
+    pub method: BytecodeMethod<'gc>,
 
     /// Closure scope stack at time of creation
     pub scope: Option<GcCell<'gc, Scope<'gc>>>,
@@ -191,12 +38,12 @@ pub enum Executable<'gc> {
     /// Code defined in Ruffle's binary.
     ///
     /// The second parameter stores the bound reciever for this function.
-    Native(NativeFunction<'gc>, Option<Object<'gc>>),
+    Native(NativeMethod<'gc>, Option<Object<'gc>>),
 
     /// Code defined in a loaded ABC file.
     Action {
         /// The method code to execute from a given ABC file.
-        method: Avm2MethodEntry<'gc>,
+        method: BytecodeMethod<'gc>,
 
         /// The scope stack to pull variables from.
         scope: Option<GcCell<'gc, Scope<'gc>>>,
@@ -424,7 +271,7 @@ impl<'gc> FunctionObject<'gc> {
     /// Construct a builtin function object from a Rust function.
     pub fn from_builtin(
         mc: MutationContext<'gc, '_>,
-        nf: NativeFunction<'gc>,
+        nf: NativeMethod<'gc>,
         fn_proto: Object<'gc>,
     ) -> Object<'gc> {
         FunctionObject(GcCell::allocate(
@@ -440,7 +287,7 @@ impl<'gc> FunctionObject<'gc> {
     /// Construct a builtin type from a Rust constructor and prototype.
     pub fn from_builtin_constr(
         mc: MutationContext<'gc, '_>,
-        constr: NativeFunction<'gc>,
+        constr: NativeMethod<'gc>,
         mut prototype: Object<'gc>,
         fn_proto: Object<'gc>,
     ) -> Result<Object<'gc>, Error> {
