+++ core/src/avm2/script.rs
@@ -1,7 +1,7 @@
 //! Whole script representation
 
 use crate::avm2::class::Class;
-use crate::avm2::function::{Avm2MethodEntry, Method};
+use crate::avm2::method::{BytecodeMethod, Method};
 use crate::avm2::r#trait::Trait;
 use crate::avm2::Error;
 use gc_arena::{Collect, GcCell, MutationContext};
@@ -77,8 +77,8 @@ impl<'gc> TranslationUnit<'gc> {
 
         drop(write);
 
-        let method: Result<Avm2MethodEntry<'gc>, Error> =
-            Avm2MethodEntry::from_method_index(self, Index::new(method_index))
+        let method: Result<BytecodeMethod<'gc>, Error> =
+            BytecodeMethod::from_method_index(self, Index::new(method_index))
                 .ok_or_else(|| "Method index does not exist".into());
         let method: Method<'gc> = method?.into();
